requirements.txt
python-dotenv==1.0.1
requests==2.32.3       
fastapi==0.115.3        
uvicorn==0.32.0       
pytest==8.3.3        
httpx==0.27.2         
pytest-asyncio==0.24.0  